test: prefer small per-op spec files with shared beforeAll over one big meta file
Replaced the single mutation-denial-matrix.meta.spec.ts with 13 small
per-op files (matching the original convention), plus a reusable
`usingSharedProject()` scenario helper and a
`PROJECT_MUTATION_DENIAL_MATRIX` constant.

Each restored file now:
  - uses `usingSharedProject()` (beforeAll, one prepareData per file)
  - runs only `PROJECT_MUTATION_DENIAL_MATRIX` (cross-owner + anonymous)

Trade vs the big meta file: +13 × ~30ms prepareData = ~0.4s CPU,
-1 big ugly file, +13 small focused files matching the rest of the
project's convention.
