+++ src/testing/kit/auth-permission/index.ts
@@ -20,6 +20,7 @@ export {
 } from './run-auth-matrix';
 export {
   PROJECT_MUTATION_MATRIX,
+  PROJECT_MUTATION_DENIAL_MATRIX,
   ORG_MUTATION_MATRIX,
   PROJECT_VISIBILITY_MATRIX,
   PROJECT_PII_READ_MATRIX,
